// src/pages/Dashboard/components/CreateHotel/schemas/create-schema.ts
import * as yup from 'yup';

export const CreateFormSchema = yup.object({
  image: yup.mixed().required('Imagen es requerida'),
  name: yup.string().required('Nombre de Usuario es requerido'),
  place: yup.string().required('Ciudad es requerida'),
  address: yup.string().required('Dirección es requerida'),
  description: yup.string().required('Descripción es requerida')
});